utils/browser_factory.py
from selenium import webdriver
from selenium.webdriver.chrome.service import Service
from selenium.webdriver.chrome.options import Options
from webdriver_manager.chrome import ChromeDriverManager
import yaml

def create_browser():
    with open("config/config.yaml", "r") as f:
        config = yaml.safe_load(f)

    options = Options()
    if config.get("headless", False):
        options.add_argument("--headless")

    service = Service(ChromeDriverManager().install())
    driver = webdriver.Chrome(service=service, options=options)
    driver.maximize_window()
    return driver
