requirements.txt
selenium
pytest
pyyaml
pytest-html